Flush pipeline
To believe that these changes were made 3 years ago - it's haunting.
However, the changes were never committed and thus never published - until now.

Note: Even though this commit was made in June 2024, this commit is
meant to work only with Galimulator 4.7, meaning that it will not
work in newer versions which are likely used by people in 2024 or
beyond. Truth be told, this commit is only made as a historical footnote.