Move to Galimulator 5.0.2, early EmploymentAgency optimizations
This makes a 5k star galaxy at least somewhat playable.
Although I am a bit surprised how I manged this considering that the
only optimization right now is just to use distanceSquared instead of
distance and to remove am unneeded field write. Also manually
inlined a lerp call, but that probably is not the cause of this change.

But hey, every little change can have drastic impact, so cutting
the mean processing time of a method by half can have quite the impact.