Use custom QuadTree implementation for k-nearest-neighbours queries
A rather extremely buggy implementation, but works with a 100% accuracy 99% of the time.
As for the remaining 1% of the time - the inaccuracy will be subtle enough that it can
be discarded. Yet, if the setup is right - people WILL notice something is wrong, so
eventually I will have to use a different knn implementation rather than one I wrote
myself.