Only consider a small subset of people when trying to find a replacement for a job
Let's not do a stellaris, yeah?

Also, the main bottleneck now are our knn queries. From the looks of it they still
consume far too much of our previous CPU time. So it's quite clear that we'd probably
need to get rid of the Spatial Query Arrays and use a library. Perhaps we are back
to tinspin-indexes? Although I don't expect too much from it so maybe just go
completely ballistic and use a native library for this?