Try reduce amount of allocations in EmploymentAgency
This actually doesn't seem to have a notable effect, but I think
this commit does lay the groundwork for more improvements later on.

Also, a few components seem to be highly buggy at this point,
but it's all not easy to reproduce. I should look into it more
closely sooner than later.