Radically improve performance of the first EmploymentAgency tick during world gen
This now brings down the time it takes for the first tick to a comfortable
60-70 seconds - for a 250 000 star galaxy. There is probably a fair bit that
can still be optimized (ideally we'd use the tinspin indices, but they require
a version of Java that isn't suitable for galimulator modding - yet).
The knn queries take the longest amount of time - perhaps there might be a few
tricks to pull out still; My guess is that we can/should make use of vertically
spliced SpatialQueryArrays, which should have an impact especially on gigantic
maps. That being said, this remains a theory so the final performance impact
of such a solution remains unmeasured.